requirements.txt
openai==1.45.0
python-dotenv==1.0.1
anthropic==0.34.2
